Change GlobalMonitor totalRamUsage to return kbytes
<==== Mercurial data ====>
Branch: vms_4.1

+++ vms/server/nx_vms_server/src/platform/monitoring/monitor_linux.h
@@ -20,7 +20,7 @@ public:
     virtual ~QnLinuxMonitor();
 
     virtual qreal totalCpuUsage() override;
-    virtual qreal totalRamUsage() override;
+    virtual quint64 totalRamUsage() override;
     virtual QList<HddLoad> totalHddLoad() override;
     virtual QList<QnPlatformMonitor::NetworkLoad> totalNetworkLoad() override;
     virtual QList<PartitionSpace> totalPartitionSpaceInfo() override;
