+++ vms/server/nx_vms_server/src/platform/monitoring/sigar_monitor.h
@@ -15,7 +15,7 @@ public:
 
     virtual qreal totalCpuUsage() override;
 
-    virtual qreal totalRamUsage() override;
+    virtual quint64 totalRamUsage() override;
 
     virtual QList<HddLoad> totalHddLoad() override;
